Change varnishkafka to use the new VUT API
Varnish 5.2 introduced a new API to make things easier for applications that want to interact with VSL and VSM.

So switch to using this new API by removing VSM/VSL initialization from varnishkafka codebase and using the VUT entrypoints.
Main loop is in VUT so use the idle callback for varnishkafka flushing.
Setup of interrupt handlers also goes through VUT, and they need to ack the interrupts to VUT.
Use the "varnish way" to parse CLI arguments by creating a header file that describes all CLI arguments taken by the application and having a global object created to store the parsing of the args.

Change-Id: Ic7204843321f1ec3c2a54241a693b26dd0d350a7